Fixed bug with App Crash
