+++ MazesAndMinotaurs/app/src/main/java/com/example/cis/mazeminotaurs/character/classes/Centaur.java
@@ -167,6 +167,15 @@ public class Centaur extends Warrior implements Level{
         mScoreLevelChoice = scoreLevelChoice;
     }
 
+    public void setWeaponOfChoice(Weapon weaponOfChoice) {
+        if (Arrays.asList(mPossibleWeaponsOfChoice).contains(weaponOfChoice)) {
+            mWeaponOfChoice = weaponOfChoice;
+        } else {
+            System.out.println("Invalid assignment of weaponOfChoice. Assigning default.");
+            mWeaponOfChoice = mPossibleWeaponsOfChoice[0];
+        }
+    }
+
     public int getExtraoridnaryAgilityBonus(){
         return getCharacter().getScore(Score.SKILL).getModifier();
     }
