+++ MazesAndMinotaurs/app/src/main/java/com/example/cis/mazeminotaurs/character/classes/Hunter.java
@@ -11,6 +11,7 @@ import com.example.cis.mazeminotaurs.character.stats.Score;
 import com.example.cis.mazeminotaurs.util.Util;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 
@@ -19,6 +20,7 @@ import java.util.HashMap;
  */
 
 public class Hunter extends Specialist implements Level{
+    // Null for this value means that it is linked to weapon of choice.
     private Weapon[] mPossibleStartingWeapons = null;
 
     private Weapon[] mPossibleWeaponsOfChoice = new Weapon[] {
@@ -157,5 +159,14 @@ public class Hunter extends Specialist implements Level{
 
     public void setScoreLevelChoice(ArrayList<HashMap<Score, Integer>> scoreLevelChoice) {
         mScoreLevelChoice = scoreLevelChoice;
+    }
+
+    public void setWeaponOfChoice(Weapon weaponOfChoice) {
+        if (Arrays.asList(mPossibleWeaponsOfChoice).contains(weaponOfChoice)) {
+            mWeaponOfChoice = weaponOfChoice;
+        } else {
+            System.out.println("Invalid assignment of weaponOfChoice. Assigning default.");
+            mWeaponOfChoice = mPossibleWeaponsOfChoice[0];
+        }
     }
 }
